Stop using Message.user_number for finding/upserting members
For Twilio at least, this is a generated incoming number, and will
possibly be different for each campaign contact. If message.user is
nil, we simply don't know which Spoke backend account was involved.
That should never be the case for outging messages - i.e. messages
the user sends to a campaign contact, but seems to be always the
case for incoming messages - i.e. replies from the campaign contact
to a message.

Since Contact.contactee cannot be nil (but Contact.contactor can
be) it means that the Contact.contactee must always be the Spoke
campaign contact, and the Contact.contactor must be the Spoke
user given by message.user (if known), even for incoming messages.
This sits a bit awkwardly, but if you think about contactor/contactee
as being the people originating and receiving messages for the
campaign as a whole (i.e. not from a per-message perspective) it
makes a bit of sense.

 Note also that since conversations can be re-assigned, we can't
just rely on "whoever sent the first message" to work out who
 will pick up and reply to any incoming messages, either. Best
 not make a guess that may well be in correct.